🎉(repository) take charge of the repository for the openfun fork
We are forking `subsrt` to improve on some issues we encountered
with the library. We intend to publish a package from this fork and
generally maintain it and keep it up-to-date.

To this end we're making some changes to take charge of this
repository, while still prominently crediting the original author.