🐛(VTT) use existing caption.index when building vtt captions
Captions have a property index allowing the caption position. With this
info it is not needed anymore to compute this value based on a for loop
index. Moreover thie computed position was wrong because caption of type
meta does not have index and are not included to compute index position.