✨(parsing) remove byte-order-marks at the beginning of content
Byte-order-marks break parsing as it is implemented in this library.
When a developer uses a naïve approach and simple opens a file, reads
it and converts it to a string, a byte-order-mark will still be
present in the string (if there was one in the file), and potentially
break parsing and/or result in broken outputs.

One could argue stripping the BOM is a responsibility of the
developer reading the file. In our opinion, we don't see potential
scenarios where we could break anything by removing it brefore
parsing in this library, and it makes for a more robust and vastly
better experience, if it can help other users of `subsrt` avoid
spending hours tracking this down.

Hence, we're just removing `\uFEFF` from contents before parsing,
but only if it's the first character in the content (a forgotten
BOM).